fix(basic-angular-blog-app): refactor the project structure into feature based
