+++ src/app/blog/blog-content/blog-content.component.scss
