+++ src/app/blog/blog-list/blog-list.component.scss
