+++ src/app/blog/blog.selector.ts
@@ -1,7 +1,7 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { adapter, IBlogState } from '../reducers/blog.reducer';
-import { IBlog } from '../../shared/models/blog.model';
-import { BLOGS_STATE_NAME } from '../../shared/constants/state.constant';
+import { adapter, IBlogState } from './blog.reducer';
+import { IBlog } from '../shared/models/blog.model';
+import { BLOGS_STATE_NAME } from '../shared/constants/state.constant';
 
 export const selectBlogState =
   createFeatureSelector<IBlogState>(BLOGS_STATE_NAME);
