+++ src/app/discover/discover.component.html
