+++ src/app/discover/discover.component.spec.ts
