+++ src/app/discover/file-upload/file-upload.component.spec.ts
