+++ src/app/discover/map-controls/coordinates.control.ts
