+++ src/app/discover/map-controls/legend.control.ts
