+++ src/app/home/home.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { BlogListComponent } from '../blog/blog-list/blog-list.component';
+import { RouterLink } from '@angular/router';
+
+@Component({
+  selector: 'app-home',
+  standalone: true,
+  imports: [BlogListComponent, RouterLink],
+  templateUrl: './home.component.html',
+  styleUrl: './home.component.scss',
+})
+export class HomeComponent {}
