+++ src/app/shared/ui/footer/footer.component.scss
