+++ src/app/shared/ui/footer/footer.component.spec.ts
