+++ src/app/shared/ui/header/header.component.scss
